Made sure that the SHD pin OUTPUT was made high before being initialised
I noticed that the coil was powering for a short amount of time even when in sleep mode. This was down to the SHD pin of the RFID IC being pulled low by esp making the connected pin an output. Making sure this is high before doing so seems to fix the problem